+++ CHANGELOG.md
@@ -10,11 +10,13 @@
 - Add ig.ENTITY.WaveTeleport remote state
 - Add ig.ENTITY.RotateBlocker remote state
 - Add ig.ENTITY.PushPullDest remote state
+- Add ig.ENTITY.Compressor remote state
 
 ### Fixed
 
 - Fix crash on wave ball throw into ig.ENTITY.WaveTeleport
 - Disable push interact on push blocks after they enter their destination
+- Fix crash when ig.ENTITY.Compressor has followCamera flag on
 
 ## [0.8.6] 2026-09-09